First pass at simple BLE Central and Peripheral apps
The Peripheral app just starts advertising a 1Hz counter on a custom CBUUID.  Tap "Start Advertising" to start publishing it to CoreBluetooth.
The Central app will sit there until you tap "connect", and then listen for the service advertised by the Peripheral app.  It will then display the number as it changes.  Disconnect to stop.

One additional thing: The Central app has a feature where if you tap the counter, it changes its color and saves it in UserDefaults.  The idea is: if UserDefaults is ever blown away by awakening before first unlock, the color will go away.